backend/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
nltk==3.8.1
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
textblob==0.17.1
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
